FST-108 refactor and structure
